Replace BigInteger with Integer for primary key columns across models. Add the development seeding script.

+++ app/models/admin.py
@@ -2,14 +2,14 @@ from __future__ import annotations
 from datetime import datetime
 from typing import Optional
 from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy import BigInteger, String, DateTime
+from sqlalchemy import Integer, String, DateTime
 from . import Base
 
 
 class Admin(Base):
     __tablename__ = "admins"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     paymail: Mapped[str] = mapped_column(String(100), unique=True, nullable=False)
     password_hash: Mapped[str] = mapped_column(String(255), nullable=False)
     name: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
